benchmark iteration

+++ canary/benchmark.sh
@@ -16,7 +16,7 @@ mtu=9001
 echo Enumerating local devices...
 devices=($(ip link show | grep -E '^[0-9]+:[ ]+eth' | sed -E 's/[0-9]+: eth([0-9]+).+/eth\1/'))
 
-config=canary_config_no_upload_100.json
+config=../canary_config_no_upload_100.json
 
 while (( "$#" )); do
 	case "$1" in
@@ -48,8 +48,8 @@ while (( "$#" )); do
             echo CLI Devices: ${devices[@]}
             shift
             ;;
-        --upload)
-            config_file=canary_config_upload_100.json
+        --config)
+            config_file=$(echo $1 | cut -f2 -d=)
             shift
             ;;
 		*)
@@ -58,6 +58,11 @@ while (( "$#" )); do
 			;;
 	esac
 done
+
+if [ ! -e "$config_file"; ]; then
+    echo Config file ${config_file} does not exist
+    exit 1
+fi
 
 rm -f /tmp/benchmark_*.log
 
@@ -82,16 +87,14 @@ sudo sysctl -w net.core.netdev_max_backlog=$backlog
 declare -a pids=()
 idx=0
 for local_ip in ${local_ips[*]}; do
-	echo Launching on ${devices[${idx}]}/${local_ip}
+	echo -n Launching on ${devices[${idx}]}/${local_ip}...
 	numa_node=${numa_nodes[$idx]}
 	if [ -n "$numactl" ] && [ -n "$numa_node" ]; then
 		numactl="${numactl} --${numactl_mode}=${numa_node}"
 	fi
     log_file=/tmp/benchmark_${devices[$idx]}.log
 
-	cmd="LD_PRELOAD=${bindhack} BIND_SRC=${local_ip} ${numactl} ../build/canary/aws-crt-cpp-canary -g ../${config_file} 2>&1 > ${log_file}"
-    echo $cmd
-    LD_PRELOAD=${bindhack} BIND_SRC=${local_ip} ${numactl} ../build/canary/aws-crt-cpp-canary -g ../${config_file} 2>&1 > ${log_file} &
+    LD_PRELOAD=${bindhack} BIND_SRC=${local_ip} ${numactl} ../build/canary/aws-crt-cpp-canary -g ${config_file} 2>&1 > ${log_file} &
 	pids+=($!)
 	echo Launched $!
 	idx=$(( $idx + 1 ))
